Improved Walking Animation
I updated the left and right walking movement animations so that they are more in line with the character hitbox. I also changed a few pixels to make it clearer which leg is the furthest forward one. It's still going to kook weird until I get around to adding two more frames to the walking animations.